Use geopy for getting uw lat & lng
Instead of using the Google Maps API, use the geopy lib with Open Cage. This is
better for bulk transactions and handling the getting of lat and lng.

This code also handles the scenarios where the zip code is not set, which is not
a very important aspect of setting the location. If there is no zip, then it
will not be rendered.

The same goes for name - a UW must have a name.